+++ walk/randomwalk_test.go
@@ -5,6 +5,8 @@ import (
 	"github.com/stretchr/testify/assert"
 	"math/rand"
 	"time"
+	"github.com/google/gofuzz"
+	"log"
 )
 
 func TestConstructorSetsVals(t *testing.T) {
@@ -55,7 +57,7 @@ func TestInterfaceCallsMethodsEqually(t *testing.T) {
 }
 
 func TestNilWalkerReturnsError(t *testing.T) {
-	assert.Error(t, New(nil).Walk(32), "error not thrown")
+	assert.EqualError(t, New(nil).Walk(32), "Walker implemenation not set")
 }
 
 func TestMaxWeights(t *testing.T) {
@@ -166,13 +168,35 @@ func TestZeroWeightsReturnErrorWalking(t *testing.T) {
 	mw := &TestWalker{}
 	rw := NewRandomWalk(0, 0, 0, 0, mw)
 	err := rw.Walk(maxIts)
-	assert.Error(t, err)
-	assert.Equal(t, "Sum of weights is <= 0", err.Error())
+
+	assert.EqualError(t, err, "Sum of weights is <= 0", err.Error())
 
 	assert.Equal(t, 0, mw.l, "left sums not 0")
 	assert.Equal(t, 0, mw.r, "right sums not 0")
 	assert.Equal(t, 0, mw.u, "up sums not 0")
 	assert.Equal(t, 0, mw.d, "down sums not 0")
+}
+
+func TestWalkWithFuzz(t *testing.T) {
+	var maxIts uint32
+	fuzz.New().NilChance(0).Funcs(func(e *uint32, c fuzz.Continue){
+		*e = uint32(rand.New(rand.NewSource(time.Now().UnixNano())).Int31n(5000000) + 10000)
+	}).Fuzz(&maxIts)
+
+	mw := &TestWalker{}
+	var rw RandomWalk
+	fuzz.New().NilChance(0).Funcs(func(e *RandomWalk, c fuzz.Continue) {
+		c.Fuzz(&e.Left)
+		c.Fuzz(&e.Right)
+		c.Fuzz(&e.Up)
+		c.Fuzz(&e.Down)
+	}).Fuzz(&rw)
+
+	rw.Walker = mw
+
+	err := rw.Walk(maxIts)
+	assert.NoError(t, err, "", rw)
+	log.Println(rw, maxIts)
 }
 
 type TestWalker struct {
@@ -195,7 +219,8 @@ func (w *TestWalker) Down() {
 	w.d++
 }
 
-func genWeights() (Weight, Weight, Weight, Weight) {
+func
+genWeights() (Weight, Weight, Weight, Weight) {
 	rng := rand.New(rand.NewSource(time.Now().UnixNano()))
 	return Weight(rng.Uint32()), Weight(rng.Uint32()), Weight(rng.Uint32()), Weight(rng.Uint32())
 }