+++ CHANGELOG.md
@@ -6,6 +6,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Testing with gofuzz
+
+
 ## [1.0.1] 12-03-2017
 ### Added
 - Bug fix to return error when no weights have been set
@@ -17,5 +22,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 [1.0.0]: https://github.com/while-loop/go-walk/releases/tag/1.0.0
 [1.0.1]: https://github.com/while-loop/go-walk/compare/1.0.0...1.0.1
+[Unreleased]: https://github.com/while-loop/go-walk/compare/1.0.1...develop
 
 [comment]: # (Added, Changed, Removed)